Pull sorting functions into one file
